클래스 파일 분할 게임 클래스의 하위로 유저번호, 컴퓨터번호 클래스 이동
