+++ docs/README.md
@@ -1,4 +1,5 @@
-# 사용자숫자
+# 기능명세서
+## 사용자숫자
      [v] 사용자로부터 게임 숫자 입력
      [v] 저장된 숫자 불러오기
      [v] 입력한 문자열을 List<Integer>로 변환
@@ -8,11 +9,11 @@
      [v] 숫자형식 검사
      [v] 1~9 검사
 
-# 컴퓨터숫자
+## 컴퓨터숫자
      [v] 새로운 숫자 생성
      [v] 저장된 숫자 불러오기
 
-# 게임
+## 게임
      [v] 게임 시작 문자열 출력
      [v] 숫자 입력 요청 문자열 출력
      [v] 게임 재시작 문자열 출력
